+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/export/PlainTextFileBasedExportRoomSinkConverter.kt
@@ -50,7 +50,7 @@ private val exportTimestampMessageFormat by lazy {
         chars("-")
         monthNumber()
         chars("-")
-        dayOfMonth()
+        day()
         chars(" ")
         hour()
         chars(":")